+++ insts/std/MOVE/setup.h
@@ -1,4 +1,4 @@
 /* setup.h */
-int get_setup_params(double [], double [12][12], float *, float *, int *,
+int get_setup_params(double [], double [12][12], double *, double *, int *,
                                                          double *, double *);
